+++ src/main/java/net/zhaiji/chestcavitybeyond/api/event/OrganRegisterEvent.java
@@ -1,0 +1,10 @@
+package net.zhaiji.chestcavitybeyond.api.event;
+
+import net.neoforged.bus.api.Event;
+import net.neoforged.fml.event.IModBusEvent;
+
+/**
+ * 器官注册事件
+ */
+public class OrganRegisterEvent extends Event implements IModBusEvent {
+}